backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
aiofiles==23.2.1
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.26.0
aiohttp==3.9.1
